Allow query based on multiple item ids.
This uses a similar flow to that of recommendations from a user id.  The process is to use a user id, fetch events for it, and base recommendations on those event items.  This change allows the client to pass in event ids to use as events.  To use it, call the API with "userItems" and an array of item ids as the value.

An alternative implementation might have been to make a variant of the "item" query and take "items", but the logic involved in item finds similar items to the input item, where running that for n items might yield unexpected results without handling the "items" together (if they were just simply to be independantly searched for similarity and then combined).